+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.0] - 2026-01-14
+
+- implementation for r2c, c2r, c2c for fftw and cufft
+- implementation for in-place transforms
+- testing validated for cpu and gpu
+
 ## [0.0.2] - 2026-01-14
 
 - integrated serial binding for fftw and cufft